Fix non-random Fseed bug
Rather than storing the secret seed in pk, we overwrote the
secret seed with uninitialized memory that was then interpreted
as the seed. This results in non-uniformly random F.

This did not immediately lead to a vulnerability, but would have made
multi-target attacks against multiple public keys easier, as the
chance of repeated Fseed values increased dramatically.